Throughput error
Value of startTs is set to 0.

+++ flink-monitor/src/main/java/it/uniroma2/controllers/flink/Preprocess.java
@@ -10,7 +10,7 @@ import java.util.Comparator;
 
 public class Preprocess extends AbstractQuery<RESTBatchResponse> {
 
-    public Preprocess(DataStream<RESTBatchResponse> inputStream, double startTs) {
+    public Preprocess(DataStream<RESTBatchResponse> inputStream, long startTs) {
         super(inputStream, startTs);
     }
