+++ flink-monitor/src/main/java/it/uniroma2/controllers/flink/Query3.java
@@ -16,7 +16,7 @@ public class Query3 extends AbstractQuery<TileQ2> {
     public static int EPS = 20;
     public static int MIN_POINTS = 4;
 
-    public Query3(DataStream<TileQ2> inputStream, double startTs) {
+    public Query3(DataStream<TileQ2> inputStream, long startTs) {
         super(inputStream, startTs);
     }
