+++ flink-monitor/src/main/java/it/uniroma2/entities/query/TileQ2.java
@@ -8,7 +8,7 @@ public class TileQ2 extends TileQ1 {
 
     protected List<Outlier> outliers;
 
-    public TileQ2(int size, String printID, int seqID, int layerID, int tileID, int[][] values, double arrivalTs, int saturatedPoints, List<Outlier> outliers) {
+    public TileQ2(int size, String printID, int seqID, int layerID, int tileID, int[][] values, long arrivalTs, int saturatedPoints, List<Outlier> outliers) {
         super(size, printID, seqID, layerID, tileID, values, arrivalTs, saturatedPoints);
         this.outliers = outliers;
     }
