+++ flink-monitor/src/main/java/it/uniroma2/entities/query/TileQ3.java
@@ -11,7 +11,7 @@ public class TileQ3 extends TileQ2 {
         return centroids;
     }
 
-    public TileQ3(int size, String printID, int seqID, int layerID, int tileID, int[][] values, double arrivalTs, int saturatedPoints, List<Outlier> outliers, List<Centroid> centroids) {
+    public TileQ3(int size, String printID, int seqID, int layerID, int tileID, int[][] values, long arrivalTs, int saturatedPoints, List<Outlier> outliers, List<Centroid> centroids) {
         super(size, printID, seqID, layerID, tileID, values, arrivalTs, saturatedPoints, outliers);
         this.centroids = centroids;
     }
